Feat: 주차권 상세 페이지 거래 기능 구현 - API에서 받는 Response에 주차장 주소가 없음 수정 필요

+++ AOS/app/src/main/java/com/team/parking/data/model/ticket/TicketBoughtListResponse.kt
@@ -8,6 +8,6 @@ data class TicketBoughtListResponse(
     val parkingRegion: String,
     val ptToLose: Int,
     val sellConfirm: Boolean,
-    val shaId: Int,
-    val ticketId: Int
+    val shaId: Long,
+    val ticketId: Long
 )
